% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stere-grid.R
\name{stere_grid}
\alias{stere_grid}
\title{Stereographic south polar grid}
\usage{
stere_grid(res = 24)
}
\description{
Generate a grid at one of a set of common resolutions. The input `res` is the resolution
in kilometres. The grid is carefully aligned to `x = 0` so that cell edges intersect exactly 
with the prime meridan and the dateline.
}
\details{
Currently `res` must be one of 24, 16, 8, 4, 1.
}
\examples{
stere_grid()
stere_grid(16)
}
